app: models: Add associations between comments and posts.
The dictionary (hash) `{:dependent => :destroy}` serves the purpose of
destroying all the comments whenever the post they belong to gets
`destroy`ed.  As an example, with these options passed to has_many, the
following code (in `rails console`):

    p = Posts.all
    p[0].destroy

would destroy all the comments associated (via their `post_id`) with the
post at the first position of the array `p`.