views: posts: Refactor to use code from partial template.
While we are at it, we also reverse the order of the displayed comments, so
that they appear automatically in reverse chronological order, which matches
the order that results when we prepend comments to a post, via the
ajax-ified version of the addition of comments.

Note that while we are removing the code from the *posts* view, the
refactored code that will be called will, from now on, be the code from the
*comments* view.

Note: It is not exactly clear to me *what* in the code makes the code from
the comments view be called from the posts view.